requirements.txt
cachetools==6.2.0
catboost==1.2.8
contourpy==1.3.2
cuda-bindings==12.9.2
cuda-pathfinder==1.2.2
cuda-python==12.9.2
cudf-cu12==25.8.0
cupy-cuda12x==13.6.0
cycler==0.12.1
fastrlock==0.8.3
fonttools==4.59.2
fsspec==2025.9.0
graphviz==0.21
joblib==1.5.2
kiwisolver==1.4.9
libcudf-cu12==25.8.0
libkvikio-cu12==25.8.0
librmm-cu12==25.8.0
lightgbm==4.6.0
llvmlite==0.44.0
markdown-it-py==4.0.0
matplotlib==3.10.6
mdurl==0.1.2
narwhals==2.4.0
numba==0.61.2
numba-cuda==0.14.1
numpy==2.2.6
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-nccl-cu12==2.28.3
nvtx==0.2.13
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==23.0.1
plotly==6.3.0
psutil==7.1.0
pyarrow==19.0.1
Pygments==2.19.2
pylibcudf-cu12==25.8.0
pynvjitlink-cu12==0.7.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rapids-logger==0.1.1
rich==14.1.0
rmm-cu12==25.8.0
scikit-learn==1.7.2
scipy==1.15.3
setuptools==65.5.1
six==1.17.0
sns==0.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.44.0
xgboost==3.0.5
